Change version to 2.0.0
It looks like everything in the projects tab is done code-wise. It only makes sense to set the version to 2.0.0 now, but I'm opening a PR just in case you had anything else you wanted to edit. Once this is approved, I'll work on updating crates.io and the prebuilt binary.